+++ NEWS.md
@@ -7,6 +7,7 @@
 * switched to {maps} package in base R advanced vignette
 * added filter (license) argument to get_uuid, pick_phylo, add_phylopic_base, add_phylopic, and geom_phylopic (#72)
 * added img argument to get_uuid and get_attribution
+* added plot and print methods for silhouette objects (#73)
 
 # rphylopic 1.1.1